+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Lives left are now shown as a title instead of the chat
+
 ## 1.0.1
 * Fixed "Half hearts' penalty on death" could not be zero or negative